refactor: move AppleSiliconMeasure to where it could be shared
